Change WIN payout to be always proportional

+++ trxbetbot/plugins/bet/bet.py
@@ -16,9 +16,7 @@ from trxbetbot.tronscan import Tronscan
 from ..autobet.autobet import Autobet
 
 
-# TODO: Handle cases:
-#  1) If balance not sufficient, stop.
-#  2) If no own wallet and user disabled or blocked bot, stop
+# TODO: Autobet: If user has bot wallet and disabled or blocked bot, stop
 class Bet(TrxBetBotPlugin):
 
     _WON_DIR = "won"
@@ -751,7 +749,7 @@ class Bet(TrxBetBotPlugin):
                         # We found the data that we will use to determine WIN amount to pay
                         if bonus_month == current_month and bonus_year == current_year:
                             logging.info(f"Job {bet_addr58} - 1 WIN per {trx} TRX")
-                            win_to_pay = int(amo / float(trx))
+                            win_to_pay = amo / float(trx)
 
                             if win_to_pay > 0:
                                 sent_win = TRC20().send("WIN", self.get_tron(), bet.usr_address, win_to_pay)
